requirements.txt
Django==1.9
django-cors-headers==1.2.0
django-material==0.8.0
djangorestframework==3.4.7
Django==1.9
django-cors-headers==1.2.0
django-material==0.8.0
djangorestframework==3.4.7
mysqlclient==1.3.9
